feat(core): switch to a require script (node -r) because signals are hard (#42)
* feat(core): switch to a require script (node -r) because signals are hard

signals are tough, and yarn pnp does not support bash scripts. nodejs does not give the ability for
a script to replace its own process, meaning signals would need to be manually forwarded. It was
easier to skip the problem completely by using node's `-r` ability.

BREAKING CHANGE: Previously `ts-run` was an 'npm bin'. Now it is a bootstrap script: node -r
@kirbysayshi/ts-run ./path/to/your/script.ts

* chore(build): fix deps

* test: remove unnecessary test